Sum lead metrics across pages in multipart guides.
In multipart guides, we want our lead metrics to show the number
of pageviews, problem reports etc across all pages in the guide.
This is more intuitive for users than showing an average of an
average.

To this end, add new methods on our Metrics class to return
these numbers.

These methods work by dividing the sum of all pageviews (or other
metric) by the number of days in the time period queried. We calculate
this by looking at the number of results returned for a particular
URL.

Note that this assumes the PP always returns a data entry for each day
across the time period queried, even if the data value is null or zero.
This assumption is already made elsewhere in the application.

We render these numbers in the template when the format is 'guide'.
In future we may want to use these methods for smart answers too.

Finally, update tests to mimic the real data returned by the Performance
Platform.