Improve handling of multipart formats.
Two major behind-the-scenes improvements:

- Pass boolean "is_multipart" parameter to templates, rather than
  string "format". This simplifies template code.
- Create new MultiPartMetrics class that inherits from existing
  Metrics class. Move division logic out of controller and into
  class. This simplifies both template and controller code.

This MultiPartMetrics class will return averages across days, not
the average of all values given. So for example, it will return
the average problem reports across all parts of a multi-part
guide by day, not the average of an average.

Create new test for MultiPartMetrics class, and update existing
tests for new template logic.