Add support for smart answers: minor template fix.
Treat smart answers as a multi-part format. This means that the
lead metrics will be averaged across all URLs in the smart answer,
and the templates will make it clear that this is a multi-part
format.

Also removes an unused template variable and tidies template layout.

Background to smart answers logic
---------------------------------

We do not expose per-page metrics for smart answers, because the Metadata API
does not return a parts list for them. I believe this is the correct
design decision for the following reasons:

- Smart answers often generate a great many unique URLs (e.g. there are
  hundreds under /pip-checker), too many to display sensibly
- The lead metrics should contain enough information to indicate if
  there is a problem in some part of the smart answer. Content
  designers can check Feedex (which does contain per-part URLs) to
  identify what that problem is.