Upgrade Slimmer, explicitly set layout
This is a no-op change to app behaviour. The same Slimmer layout is used as be
before, but set in a simple, standardised way with slimmer_template, which is
easier to grep for than the other methods.

Slimmer 9.0.0 switched to core_layout [1] as the default layout, which is the
modern/recommend layout. Apps not yet using core_layout should be explicitly
marked as such, so we can easily identify them as needing updating.

I had a quick look at what would be required to use core_layout, and it's
actually very little, I think it would just need some tweaking of the local
grid styles.

Still, it's better to do that in a separate PR, so it can be reviewed/tested
specifically.

[1] alphagov/slimmer#136