Switch slimmer template to 'core_layout'
`core_layout` is the default layout since Slimmer 9.0.0 [1], and apps
should be using it where possible.

The advantage over `header_footer_only` is that `core_layout` sets a
base grid/container for the main content, so apps don't need to
replicate it each time.

This switches the layout, but not overriding the default, and removes
the grid/container code specific to this app.

The SASS `grid_layout` isn't needed, but was including `measurements`
which is relied on else where.

[1] https://github.com/alphagov/slimmer/pull/136