Use 'title' component
Use the component to avoid replicating page title styles locally, and
so we can roll changes out to page titles without needing to update
each app.

Although this change leaves 99% of the page as it was, it does introduce
a small visual diff; the "All metrics are recorded over the past 6 weeks"
line is pushed down slightly, as the title component has more bottom
margin.

I checked the design change with Rebecca Cottrell and she's happy with
the additional space anyway.