Uses SSL version of performance URL
As part of the process to replace metadata-api in info-frontend, we're now using 4 new API calls in gds-api-adapters in order to talk to the backdrop read API. This used to be handled by metadata-api.
However, the url path we were using for the performance endpoint was the non-secure version.

This commit replaces the performance url with the SSL enabled version (`http` -> `https`)

It also removes the `PP_DATA_OUT_ENDPOINT` from `test_helpers/statistics.rb` since it's not being used.