Formats nil values for performance stats
This is a follow up to https://github.com/alphagov/info-frontend/pull/71 and https://github.com/alphagov/info-frontend/pull/72
where we are in the process of removing metadata-api from info-frontend.

In order to do this, we have moved the api calls handled by metadata-api into gds-api-adapters. The api calls will be passed on by
gds-api-adapters to the backend read api. The response is parsed and rendered by info-frontend.

This commit will take into account the possibility that the response we receive from `backend read API` contain nil values.
This was a use case that was not taken into account before, so there is now a test to check that we can handle it.

Previously, the formatting of nil values was handled by metadata-api in here: https://github.com/alphagov/metadata-api/commit/81e2b670cd2defd94954cfc561bf929035826a8e#diff-931132c428c386579ddf11c313d82b28R220