Format performance slug before doing request
The statistics class was introduced to take over the duties from metadata-api: fetch data from the performance platform and format the data. However, it wasn't being strict about checking whether the page we're fetching performance data for has their slug starting with a "/".

It's not longer assuming the data it's receiving is correct and always makes sure to add the forward slash to the slug before sending the request to the performance platform.

The missing slash meant that for multi-part pages, the data would not be fetched correctly.