Prevent error message associated with listen gem
When trying to run the app locally we got the error"
`Could not load the 'listen' gem. Add `gem 'listen'` to the development group of your Gemfile`
This seems to be due to this line in
`config/environments/development.rb`:
`config.file_watcher = ActiveSupport::EventedFileUpdateChecker`

Since it was added by the rails 6 upgrade it doesn't seem necessary to
keep. So in this commit we have commented it out again.