Run `bundle exec rails app:update`
The new rails 6 defaults in
`config/initializers/new_framework_defaults_6_0.r`.
are explained in:
https://medium.com/@dylansreile/rails-6-0-new-framework-defaults-what-they-do-and-how-to-safely-uncomment-them-586146f371e8
It seems we can use all of them. They are required in
`config/application.rb`.

Running `bundle exec rails app:update` also created another file:
`config/initializers/content_security_policy.rb`.
This is to define an application-wide content security policy, but since
this is out of scope for this specific piece of work, we'll leave it
commented (as it was created).

Trello card: https://trello.com/c/0d6k3UdH/1798-5-upgrade-info-frontend-to-rails-6