Remove mentions of metrics from info pages
Since we retired performance platform in March 2021, we haven't had any metrics on the info pages.

I commented out the metrics in f7b2c8cf083b32e37e82dc4bc438f450044f879e with
half an idea that we might restore them by making requests directly to GA or
BigQuery or something.

I missed a couple of mentions of metrics on the pages though, which has left
them in a confusing state. Pages like
https://www.gov.uk/info/benefits-calculators show user needs, but no metrics.
They still mention metrics in the context of the heading, and in the "All
metrics are recorded over the past 6 weeks" statement.

Pages like https://www.gov.uk/info/government/get-involved are even worse,
because there are no user needs on the page, so the proportion of content that
mentions metrics which don't exist is much higher.

We should consider retiring this entire application, or reimplementing the
metrics if we think they're useful. In the meantime, it looks slightly less bad
if we're not referring to non-existent metrics on these pages.